Разделил утилитарные функции, поправил импорт этих функций, использовал метод replace в презентере
