+++ src/utils/point.js
@@ -2,17 +2,6 @@ import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration.js';
 dayjs.extend(duration);
 
-//Функция генерации случайных чисел из диапазона
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-//Функция получения случайного числа из массива
-const getRandomArrayElement = (elements) => (elements[getRandomInteger(0, elements.length - 1)]);
-
 const humanizePointDate = (date) => dayjs(date).format('HH:mm');
 const humanizeEventDate = (date) => dayjs(date).format('MMM D');
 
@@ -33,4 +22,4 @@ const getDurationDates = (dateStart, dateFinish) => {
   }
 };
 
-export { getRandomInteger, getRandomArrayElement, humanizePointDate, humanizeEventDate, getDurationDates };
+export { humanizePointDate, humanizeEventDate, getDurationDates };
